refactor: Moved cloudflareImagesBypass to sources/helpers and fixed NOVA novels summary format
